+++ packages/org/org_employee.js
@@ -10,6 +10,7 @@ if (ubaAuditPresent) {
 }
 
 me.on('update:after', updateCaptionAndLogToAudit)
+me.on('insert:before', buildAndFillFullname)
 me.on('insert:after', ubaAuditLinkUser)
 me.on('delete:before', removeLinkToUser)
 me.on('delete:after', ubaAuditLinkUserDelete)
@@ -431,4 +432,40 @@ function removeLinkToUser (ctx) {
     },
     __skipOptimisticLock: true
   })
+}
+
+/**
+ * If fullFIO and/or shortFIO didn't come from client, they should be built based on user's language and added to execParams
+ * @param ctx
+ */
+function buildAndFillFullname (ctx) {
+  const params = ctx.mParams.execParams
+  const { firstName, middleName, lastName, fullFIO, shortFIO } = params
+  const namePartsOrder = ({ firstName, middleName, lastName }) => {
+    const lastNameFirst = ['uk', 'ru', 'az', 'ka', 'uz'].includes(Session.userLang)
+    return lastNameFirst
+      ? [lastName, firstName, middleName]
+      : [firstName, middleName, lastName]
+  }
+  if (!fullFIO) {
+    const formattedFullName = namePartsOrder({ firstName, middleName, lastName })
+      .filter(value => !!value)
+      .join(' ')
+    params.fullFIO = formattedFullName
+  }
+
+  if (!shortFIO) {
+    const shortenedFirstName = firstName ? `${firstName[0].toUpperCase()}.` : ''
+    const shortenedMiddleName = middleName ? `${middleName[0].toUpperCase()}.` : ''
+
+    const formattedShortFullName = namePartsOrder({
+      firstName: shortenedFirstName,
+      middleName: shortenedMiddleName,
+      lastName
+    })
+      .filter(value => !!value)
+      .join(' ')
+
+    params.shortFIO = formattedShortFullName
+  }
 }